Add option for making .lcomm symbols common (#83)
This can be convenient with games using `-G` with older aspsx where
static vs. non-static symbols cause different code generation.

Without this, if you run into a function which refers to a symbol that
must be marked static to get the code generation correct, the only
option is to migrate .sdata/.sbss for the C file containing it.

With the --use-comm-for-lcomm option that isn't necessary and the usual
way of putting the address of the variable to symbols.txt still works.